+++ setup.py
@@ -3,5 +3,5 @@ from setuptools import setup
 setup(
     name='nnest',
     version='0.1.0',
-    packages=['nnest']
+    packages=['nnest'],
 )
